server/requirements.txt
-i https://pypi.org/simple
asn1crypto==1.3.0
azure-common==1.1.25
azure-core==1.7.0
azure-storage-blob==12.3.2; python_full_version >= '3.5.2'
boto3==1.14.22
botocore==1.17.22
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cryptography==2.9.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
docutils==0.15.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
fastapi==0.59.0
h11==0.9.0
httptools==0.1.1; sys_platform != 'win32' and sys_platform != 'cygwin' and platform_python_implementation != 'PyPy'
idna==2.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
isodate==0.6.0
isort==5.1.1
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
msrest==0.6.17
numpy==1.19.0; python_version >= '3.6'
oauthlib==3.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
oscrypto==1.2.0
pandas==1.0.5
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pycryptodomex==3.9.8; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pydantic==1.6.1; python_version >= '3.6'
pyjwt==1.7.1
pyopenssl==19.1.0
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2020.1
pyyaml==5.3.1
requests-oauthlib==1.3.0
requests==2.23.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
s3transfer==0.3.3
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
snowflake-connector-python==2.2.9
snowflake-sqlalchemy==1.2.3
sqlalchemy==1.3.18
starlette==0.13.4; python_version >= '3.6'
urllib3==1.25.9; python_version != '3.4'
uvicorn==0.11.5
uvloop==0.14.0; sys_platform != 'win32' and sys_platform != 'cygwin' and platform_python_implementation != 'PyPy'
websockets==8.1; python_full_version >= '3.6.1'
